remove onlink existing files

+++ src/AppBundle/Controller/PromoController.php
@@ -68,7 +68,7 @@ class PromoController extends Controller
                 $fileName = $fileUploader->upload($file);
                 $promo->setSignTrainer($fileName);
 
-                if ($signT != null){
+                if ($signT != null && file_exists($this->getParameter('sign_directory') . $signT)){
                     unlink($this->getParameter('sign_directory') . $signT);
                 }
             } else {
